Feat: rating updates working after object refactoring - app user still coming back null sometimes

+++ app/src/main/java/com/example/surfeillancefrontend/service/TripApiService.java
@@ -19,6 +19,6 @@ public interface TripApiService {
 
 
     @PUT("users/trips/{tripid}")
-    Call<Trip> editTripByTripId(@Path("tripid") Long tripid, @Body UpdateRatingDTO ratings);
+    Call<Trip> editTripByTripId(@Path("tripid") int tripid, @Body UpdateRatingDTO ratings);
 
 }
